notification_service/tests/requirements.txt
pytest==7.4.4
pydantic==2.5.3
pydantic_settings==2.1.0
httpx==0.26.0
psycopg2-binary==2.9.9